% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_trend.R
\name{extract_trend}
\alias{extract_trend}
\alias{extract_trend.gtrend_scraper}
\alias{extract_trend.gtrend_scraper_element}
\title{Extract Trend Data}
\usage{
extract_trend(x, ...)

\method{extract_trend}{gtrend_scraper_element}(x, ...)

\method{extract_trend}{gtrend_scraper}(x, ...)
}
\arguments{
\item{x}{A \code{gtrend_scraper} object.}

\item{\ldots}{Currently ignored.}
}
\value{
Returns a \code{data.frame} (for individual \code{gtrend_scraper} elements)
or a list of \code{data.frame} for a \code{gtrend_scraper} object.
}
\description{
Extract the \code{"trend"} from a \code{gtrend_scraper} object and renames
last column "trend".
}
\details{
The trend is not an absolute trend but a relative index.  See:
\url{https://support.google.com/trends/answer/4355164?hl=en} for more.
}
\examples{
\dontrun{
out <- gtrend_scraper("tyler.rinker@gmail.com", "password", c("cat", "dog bite"))
extract_trend(out)
extract_trend(out[[1]])
}
}

