% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gtrend_scraper}
\alias{gtrend_scraper}
\title{Grab Google Trend Data}
\usage{
gtrend_scraper(username, password, phrases, geo = "US")
}
\arguments{
\item{username}{Your Gmail/Google name (e.g., \code{tyler.rinker@gmail.com}).}

\item{password}{Your Gmail/Google password.}

\item{phrases}{A character vector of phrases to look up.}

\item{geo}{A locations to search within.}
}
\value{
Returns a list of lists (one list for each search term):
\item{query}{A character... }
\item{meta}{A character... }
\item{trend}{A data.frame... }
\item{regions}{A list... }
\item{topmetros}{A list... }
\item{cities}{A list... }
\item{searches}{A list... }
\item{rising}{A list... }
\item{headers}{A character... }
}
\description{
A function to grab \href{http://www.google.com/trends/}{Google Trend} data
using the \pkg{GTrendsR} package \url{https://github.com/dvanclev/GTrendsR}.
}
\examples{
\dontrun{
(out <- gtrend_scraper("tyler.rinker@gmail.com", "password", "dog bite"))

library(GTrendsR)
as.zoo(out[[1]])
}
}
\references{
\url{https://github.com/dvanclev/GTrendsR}
\url{http://www.google.com/trends}
}

