% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.trend2long}
\alias{plot.trend2long}
\title{Plots a trend2long Object}
\usage{
\method{plot}{trend2long}(x, size = 1, ...)
}
\arguments{
\item{x}{The trend2long object.}

\item{size}{The line size of the trend line.}

\item{\ldots}{ignored}
}
\description{
Plots a trend2long object.
}

