% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.trend_facet}
\alias{plot.trend_facet}
\title{Plots a trend_facet Object}
\usage{
\method{plot}{trend_facet}(x, rect.alpha = 0.025, ncol = 2, ...)
}
\arguments{
\item{x}{The trend_facet object.}

\item{rect.alpha}{The alpha level of the rectangles in the year plot
background.}

\item{ncol}{The number of facet columns.}

\item{\ldots}{Other arguments passed to
\code{\link[plotflow]{unbalanced_facet_axis}}.}
}
\description{
Plots a trend_facet object.
}

