% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trend2long.R
\name{trend2long}
\alias{as.trend2long}
\alias{trend2long}
\title{extract_trend List Output to Long Format}
\usage{
trend2long(x, ...)

as.trend2long(x, ...)
}
\arguments{
\item{x}{An extract_trend list output.}

\item{\ldots}{Currently ignored.}
}
\value{
returns a \code{data.frame} of trend data by lookup terms.

Returns a a \code{trend2long} object.
}
\description{
Reshape a extract_trend list output to long format with a term colun added.

Coerce a data.frame to a \code{trend2long} object.
}
\note{
There is no object checking.  It is the user's responsibility to make
sure the object being coerced correspondes to a \code{trend2long} object.
}
\examples{
\dontrun{
library(dplyr)
out <- gtrend_scraper("tyler.rinker@gmail.com", "password",
    c("ebola", "sars", "hiv", "aids", "influenza", "flu", "the cold"))

as.zoo.gtrends(out[[1]])

trend2long(out) \%>\%
   plot

library(dplyr)
trend2long(out) \%>\%
   filter(term \%in\% c("hiv", "ebola", "sars")) \%>\%
   as.trend2long \%>\%
   plot

library(ggplot2)
trend2long(out) \%>\%
    ggplot(aes(x=start, y=term, fill=trend)) +
    geom_tile() +
    scale_fill_gradient(low="black", high="white") +
    scale_x_date(expand = c(0, 0)) +
    theme_bw() + xlab("time")
}
}

