% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trend_facet}
\alias{trend_facet}
\title{Faceted Trend Plot}
\usage{
trend_facet(x, ...)
}
\arguments{
\item{x}{A \code{trend2long} or \code{gtrend_scraper} object.}

\item{\ldots}{ignored.}
}
\description{
This function creates a \code{data.frame} of the class \code{trend_facet},
which has its own plot method which is a faceted plot.
}
\examples{
\dontrun{
library(dplyr)
(out <- gtrend_scraper("tyler.rinker@gmail.com", "password", "dog bite"))

trend_facet(out) \%>\%
   plot

plotflow::unbalanced_facet_axis(trend_facet(out) \%>\%
    plot(ncol=3))

plotflow::unbalanced_facet_axis(trend_facet(out) \%>\%
    plot(rect.alpha=.4))


library(dplyr)
out <- gtrend_scraper("tyler.rinker@gmail.com", "password",
    c("ebola", "sars", "hiv", "aids", "influenza", "flu", "the cold"))

trend2long(out) \%>\%
   plot()

trend2long(out) \%>\%
   trend_facet \%>\%
   plot

out \%>\%
   trend_facet \%>\%
   plot
}
}

